ml-service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=2.0.0
joblib>=1.3.2
pydantic>=2.5.0
python-dotenv>=1.0.0

